Don't look at `.parameters` in methodshow
The parameters of `sig` are not guaranteed to match the arguments
that were passed on method construction, since the type may have
been normalized. This will become more pronounced after #31681,
but with OpaqueClosure we now have some method whose sig is
Tuple{Vararg{Any}}, so we might as well make this change now.
Of course, we may want to print OpaqueClosure methods differently
in general, but that's a question for another time, once we've
worked out exactly what the sig fields of OpaqueClosure methods
should look like.